Repair UT data for May 19 and 20
State of UT had an outage on the data site for May 19, which caused
the data for that date to be a duplicate of May 18. Repaired the
data by manually reconstructing information from local news at
this url, which was reported late in the day:

https://www.ksl.com/article/46754950/utah-reports-highest-single-day-covid-19-death-total-of-8-with-134-new-cases

Incremental case changes for quantities on May 20 were then manually adjusted to reflect
the additional cases recorded here manually for May 19.